java/src/test/java/com/yourtion/leetcode/daily/m07/d30/SolutionTest.java
package com.yourtion.leetcode.daily.m07.d30;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.MethodSource;

import java.util.stream.Stream;

import static org.junit.jupiter.params.provider.Arguments.arguments;

@DisplayName("每日一题 - 20200729")
class SolutionTest {

    static Stream<Arguments> testDataProvider() {
        return Stream.of(
                arguments(2, 1),
                arguments(10, 36)
        );
    }

    @ParameterizedTest()
    @MethodSource("testDataProvider")
    void integerBreak(int source, int res) {
        System.out.printf("runTest: %d , res: %d", source, res);
        Assertions.assertEquals(res, new Solution().integerBreak(source));
    }
}